Split services/ into services/ and foundation/
We're getting back to the point where we have a bunch of foundation APIs
and it's getting confusing having them mixed with services/.
