+++ packages/flutter/lib/foundation.dart
@@ -1,0 +1,14 @@
+// Copyright 2016 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// Core Flutter framework primitives.
+///
+/// The features defined in this library are the lowest-level utility
+/// classes and functions used by all the other layers of the Flutter
+/// framework.
+library foundation;
+
+export 'src/foundation/assertions.dart';
+export 'src/foundation/binding.dart';
+export 'src/foundation/print.dart';
