+++ packages/flutter/lib/rendering.dart
@@ -48,4 +48,5 @@ export 'src/rendering/table.dart';
 export 'src/rendering/view.dart';
 export 'src/rendering/viewport.dart';
 
+export 'package:flutter/foundation.dart';
 export 'package:vector_math/vector_math_64.dart' show Matrix4;
