+++ packages/flutter/lib/src/gestures/binding.dart
@@ -5,7 +5,7 @@
 import 'dart:typed_data';
 import 'dart:ui' as ui show window;
 
-import 'package:flutter/services.dart';
+import 'package:flutter/foundation.dart';
 import 'package:mojo/bindings.dart' as mojo_bindings;
 import 'package:mojo/core.dart' as mojo_core;
 import 'package:sky_services/pointer/pointer.mojom.dart';
