+++ packages/flutter/lib/src/rendering/binding.dart
@@ -4,6 +4,7 @@
 
 import 'dart:ui' as ui show window;
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
