+++ packages/flutter/lib/src/widgets/framework.dart
@@ -9,12 +9,12 @@ import 'dart:developer';
 import 'debug.dart';
 
 import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
+import 'package:flutter/foundation.dart';
 import 'package:meta/meta.dart';
 
 export 'dart:ui' show hashValues, hashList;
 export 'package:flutter/rendering.dart' show RenderObject, RenderBox, debugPrint;
-export 'package:flutter/services.dart' show FlutterError;
+export 'package:flutter/foundation.dart' show FlutterError;
 
 // KEYS
