+++ packages/flutter/test/gestures/gesture_binding_test.dart
@@ -4,8 +4,8 @@
 
 import 'dart:ui' as ui;
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
-import 'package:flutter/services.dart';
 import 'package:mojo/bindings.dart' as mojo_bindings;
 import 'package:sky_services/pointer/pointer.mojom.dart';
 import 'package:test/test.dart';
