+++ packages/flutter/test/gestures/gesture_tester.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/services.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 
 class TestGestureFlutterBinding extends BindingBase with Gesturer { }
