+++ packages/flutter/test/scheduler/animation_test.dart
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:flutter/services.dart';
 import 'package:test/test.dart';
 
 class TestSchedulerBinding extends BindingBase with Scheduler { }
